refactor: common 패키지 생성
