+++ src/main/java/kitchenpos/common/infra/client/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra.client;
+package kitchenpos.common.infra.client;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
