+++ src/main/java/kitchenpos/common/infra/client/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra.client;
+package kitchenpos.common.infra.client;
 
 import java.math.BigDecimal;
 import java.util.UUID;
