+++ src/main/java/kitchenpos/common/infra/client/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra.client;
+package kitchenpos.common.infra.client;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
