+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.application;
 
-import kitchenpos.infra.client.KitchenridersClient;
+import kitchenpos.common.infra.client.KitchenridersClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
